Move phase field handling to own file
