+++ src/controllers/dynakube/dynakube_controller.go
@@ -30,7 +30,6 @@ import (
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/rest"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -286,84 +285,6 @@ func (controller *DynakubeController) reconcileDynaKube(ctx context.Context, dkS
 
 	upd = controller.determineDynaKubePhase(dkState.Instance)
 	dkState.Update(upd, defaultUpdateInterval, "dynakube phase changed")
-}
-
-func (controller *DynakubeController) determineDynaKubePhase(instance *dynatracev1beta1.DynaKube) bool {
-	if instance.NeedsActiveGate() {
-		agPods, err := controller.numberOfMissingActiveGatePods(instance)
-		if err != nil {
-			log.Error(err, "activegate sts could not be accessed", "dynakube", instance.Name)
-			return updatePhaseIfChanged(instance, dynatracev1beta1.Error)
-		}
-		if agPods > 0 {
-			log.Info("activegate sts is still deploying", "dynakube", instance.Name)
-			return updatePhaseIfChanged(instance, dynatracev1beta1.Deploying)
-		}
-		if agPods < 0 {
-			log.Info("activegate sts not yet available", "dynakube", instance.Name)
-			return updatePhaseIfChanged(instance, dynatracev1beta1.Deploying)
-		}
-	}
-
-	if instance.CloudNativeFullstackMode() || instance.ClassicFullStackMode() || instance.HostMonitoringMode() {
-		oaPods, err := controller.numberOfMissingOneagentPods(instance)
-		if err != nil {
-			log.Error(err, "oneagent daemonset could not be accessed", "dynakube", instance.Name)
-			return updatePhaseIfChanged(instance, dynatracev1beta1.Error)
-		}
-		if oaPods > 0 {
-			log.Info("oneagent daemonset is still deploying", "dynakube", instance.Name)
-			return updatePhaseIfChanged(instance, dynatracev1beta1.Deploying)
-		}
-		if oaPods < 0 {
-			log.Info("oneagent daemonset not yet available", "dynakube", instance.Name)
-			return updatePhaseIfChanged(instance, dynatracev1beta1.Deploying)
-		}
-	}
-
-	return updatePhaseIfChanged(instance, dynatracev1beta1.Running)
-}
-
-func (controller *DynakubeController) numberOfMissingOneagentPods(instance *dynatracev1beta1.DynaKube) (int32, error) {
-	dsActual := &appsv1.DaemonSet{}
-	instanceName := instance.OneAgentDaemonsetName()
-	err := controller.client.Get(context.TODO(), types.NamespacedName{Name: instanceName, Namespace: instance.Namespace}, dsActual)
-
-	if k8serrors.IsNotFound(err) {
-		return -1, nil
-	}
-	if err != nil {
-		return -1, err
-	}
-	return dsActual.Status.CurrentNumberScheduled - dsActual.Status.NumberReady, nil
-}
-
-func (controller *DynakubeController) numberOfMissingActiveGatePods(instance *dynatracev1beta1.DynaKube) (int32, error) {
-	caps := generateActiveGateCapabilities(instance)
-
-	sum := int32(0)
-	found := false
-
-	for _, c := range caps {
-		stsActual := &appsv1.StatefulSet{}
-		instanceName := capability.CalculateStatefulSetName(c, instance.Name)
-		err := controller.client.Get(context.TODO(), types.NamespacedName{Name: instanceName, Namespace: instance.Namespace}, stsActual)
-
-		if k8serrors.IsNotFound(err) {
-			continue
-		}
-		if err != nil {
-			return -1, err
-		}
-		found = true
-		sum += stsActual.Status.CurrentReplicas - stsActual.Status.ReadyReplicas
-	}
-
-	if !found {
-		return -1, nil
-	}
-
-	return sum, nil
 }
 
 func updatePhaseIfChanged(instance *dynatracev1beta1.DynaKube, newPhase dynatracev1beta1.DynaKubePhaseType) bool {
